inline babel plugin into esbuild bundler
